no-mistakes(review): tests(cie): move the shared _clean_scene fixture into conftest
