+++ tests/cie/test_bin_duplicate_bone_ids.py
@@ -20,8 +20,7 @@ import bpy
 from mathutils import Vector
 import pytest
 
-from albam.lib import fs_registry
-from tests.conftest import close_new_fs_roots, remove_new_vfs_roots, vfs_root_names
+from tests.conftest import vfs_root_names
 from tests.cie.lfs_paths import resolve_archive_hashes
 from tests.cie.synthetic_bin import bone_table, build_mesh_bin
 from tests.cie.test_bin_serialization import _is_mesh_bin
@@ -47,21 +46,6 @@ BONES = [
 # the recorded ids rather than the weights the thing the table comes from.
 WEIGHTS = [((0, 0, 0), (100, 0, 0), 1), ((2, 0, 0), (100, 0, 0), 1)]
 WEIGHT_INDICES = [0, 0, 1, 1]
-
-
-@pytest.fixture
-def _clean_scene():
-    # vfs, exported and bpy.data are session-scoped state: register() runs
-    # once per pytest session, so a test that leaves objects or roots behind
-    # changes what the next one sees.
-    before = fs_registry.keys()
-    before_roots = vfs_root_names()
-    yield
-    bpy.ops.object.select_all(action="SELECT")
-    bpy.ops.object.delete(use_global=True)
-    remove_new_vfs_roots(before_roots)
-    bpy.context.scene.albam.exported.file_list.clear()
-    close_new_fs_roots(before)
 
 
 def pytest_generate_tests(metafunc):
